backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.37.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Pydantic for data validation
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.1.0

# LangChain and AI
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
langgraph==0.3.0

# Google AI
google-generativeai==0.7.2
google-cloud-aiplatform==1.71.0
google-auth==2.27.0
google-cloud-storage==2.14.0
google-cloud-tasks==2.14.2

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
pyyaml==6.0.1
Pillow==10.4.0
pandas==2.1.4
tiktoken==0.7.0
structlog==25.5.0

# OAuth
authlib==1.6.0
itsdangerous==2.1.0
